+++ worldeditadditions/init.lua
@@ -45,3 +45,4 @@ dofile(worldeditadditions.modpath.."/lib/forest.lua")
 dofile(worldeditadditions.modpath.."/lib/ellipsoidapply.lua")
 
 dofile(worldeditadditions.modpath.."/lib/subdivide.lua")
+dofile(worldeditadditions.modpath.."/lib/selection/stack.lua")