+++ CHANGELOG.md
@@ -4,6 +4,8 @@ It's about time I started a changelog! This will serve from now on as the main c
 
 ## v1.12 (unreleased)
  - Add `//srect` (_select rectangle_) - thanks, @VorTechnix!
+ - Add `//spush`, `//spop`, and `//sstack`
+
 
 ## v1.11: The big data update (25th January 2021)
  - Add `//scale` (currently **experimental**)